requirements.txt
python-telegram-bot[job-queue]==22.6
pytz==2025.2